+++ CountingGridsPy/models/__init__.py
@@ -1,3 +1,4 @@
 from .CountingGridModel import CountingGridModel
+from .CountingGridModelWithGPU import CountingGridModelWithGPU
 
-__all__ = ['CountingGridModel']
+__all__ = ['CountingGridModel','CountingGridModelWithGPU']
