+++ CountingGridsPy/tests/test_models/test_gpu.py
@@ -4,33 +4,45 @@
 import unittest
 import numpy as np
 from CountingGridsPy.models import CountingGridModel, CountingGridsModelWithGPU
+import torch
 
 
-class TestCorrectnessOfNontrivialDesignMatrix(unittest.TestCase):
+class TestGPUvsCPU(unittest.TestCase):
     def setUp(self):
-        M, N = [5, 7]
+        SEED = "03071994"
+        np.random.seed(SEED)
+        M, N = [10000, 500]
         self.N = N
-        self.data = np.array(
-            list(range(1, 8))+list(range(7, 0, -1)) +
-            [1]*7+[0, 1, 0, 1, 0, 1, 0]+list(range(1, 8))
-        ).reshape((M, N))
-
+        self.data = np.round(np.random.random((M, N)) * 10)
         self.extent = np.array([40, 40])
         self.window = np.array([5, 5])
         self.pi_init = np.ones([5]*2 + [N]) / 1000
         self.cpuModel = CountingGridModel(self.extent, self.window)
         self.gpuModel = CountingGridsModelWithGPU(self.extent, self.window)
 
-    def test_fitted_model(self):
-        self.cpuModel.fit(self.data, max_iter=50,
-                       returnSumSquareDifferencesOfPi=False, pi=np.copy(self.pi_init))
-        assert(np.all(np.isclose(self.cpuModel.q, .04)))
+    def test_fitted_model_no_layers(self):
+        numIters = 50
+        self.cpuModel.fit(
+            self.data,
+            max_iter=numIters,
+            returnSumSquareDifferencesOfPi=False,
+            pi=np.copy(self.pi_init),
+            layers=1
+        )
+
+        device = torch.device("cuda:0")
+        self.gpuModel.fit(
+            self.data,
+            max_iter=numIters,
+            pi=torch.tensor(self.pi_init, device=device, dtype=torch.double),
+            layers=1
+        )
 
         assert(
             all(
                 np.isclose(
                     self.cpuModel.q,
-                    self.gpuModel.q.numpy()
+                    self.gpuModel.q.cpu().numpy()
                 )
             )
         )
@@ -39,7 +51,44 @@ class TestCorrectnessOfNontrivialDesignMatrix(unittest.TestCase):
             all(
                 np.isclose(
                     self.cpuModel.pi,
-                    self.gpuModel.pi.numpy()
+                    self.gpuModel.pi.cpu().numpy()
                 )
             )
         )
+
+        def test_fitted_model_with_layers(self):
+            numIters = 50
+            layers = 2
+            self.cpuModel.fit(
+                self.data,
+                max_iter=numIters,
+                returnSumSquareDifferencesOfPi=False,
+                pi=np.copy(self.pi_init),
+                layers= layers
+            )
+    
+            device = torch.device("cuda:0")
+            self.gpuModel.fit(
+                self.data,
+                max_iter=numIters,
+                pi=torch.tensor(self.pi_init, device=device, dtype=torch.double),
+                layers=layers
+            )
+    
+            assert(
+                all(
+                    np.isclose(
+                        self.cpuModel.q,
+                        self.gpuModel.q.cpu().numpy()
+                    )
+                )
+            )
+    
+            assert(
+                all(
+                    np.isclose(
+                        self.cpuModel.pi,
+                        self.gpuModel.pi.cpu().numpy()
+                    )
+                )
+            )
